requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# LlamaIndex and RAG
llama-index==0.10.0
llama-index-llms-openai==0.1.5
llama-index-embeddings-openai==0.1.6
llama-index-vector-stores-chroma==0.1.4

# Vector Database
chromadb==0.4.22

# Document Processing
pypdf==4.0.0
beautifulsoup4==4.12.3
pandas==2.2.0
openpyxl==3.1.2

# OpenAI
openai==1.10.0

# Utilities
python-dotenv==1.0.1
aiofiles==23.2.1
pydantic==2.5.3
pydantic-settings==2.1.0

# HTTP Requests
httpx==0.26.0
requests==2.31.0

pytest==8.0.0
pytest-asyncio==0.23.5
httpx==0.26.0
